Added: QueryBase with CRTP to handle Query inheritance
- Added QueryBase with CRTP to handle Query inheritance and to prevent
down-casting for the children
