+++ algoliasearch-core/src/main/java/com/algolia/search/models/indexing/BrowseIndexQuery.java
@@ -1,14 +1,15 @@
 package com.algolia.search.models.indexing;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class BrowseIndexQuery extends Query {
+public class BrowseIndexQuery extends QueryBase<BrowseIndexQuery> {
 
   public BrowseIndexQuery() {}
 
   public BrowseIndexQuery(String query) {
-    super(query);
+    this.query = query;
   }
 
   public String getCursor() {
@@ -21,4 +22,10 @@ public class BrowseIndexQuery extends Query {
   }
 
   private String cursor;
+
+  @Override
+  @JsonIgnore
+  public BrowseIndexQuery getThis() {
+    return this;
+  }
 }
