+++ algoliasearch-core/src/main/java/com/algolia/search/util/QueryStringUtils.java
@@ -2,7 +2,7 @@ package com.algolia.search.util;
 
 import com.algolia.search.Defaults;
 import com.algolia.search.models.apikeys.SecuredApiKeyRestriction;
-import com.algolia.search.models.indexing.Query;
+import com.algolia.search.models.indexing.QueryBase;
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -47,7 +47,7 @@ public class QueryStringUtils {
   }
 
   @SuppressWarnings("unchecked")
-  public static String buildQueryAsQueryParams(Query query) {
+  public static String buildQueryAsQueryParams(QueryBase query) {
 
     // This could be improved
     // We need to create a Map<String, Object> to keep track of the List<List<?>>
